requirement.txt
matplotlib
numpy
opencv-python
scipy
seaborn
pandas
mtcnn
tensorflow
Flask-JWT-Extended
flask_sqlalchemy
flask-migrate
mysqlclient